backend/requirements.txt
Flask
flask-cors
transformers
torch
sentencepiece
Flask-Bcrypt
PyJWT
